requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bce-python-sdk==0.9.52
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.17.0
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.117.1
fastjsonschema==2.21.2
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.60.0
fqdn==1.5.1
fsspec==2025.9.0
future==1.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
haystack-ai==2.17.1
haystack-experimental==0.13.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
imagesize==1.4.1
imbalanced-learn==0.14.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kintera==1.1.5
kiwisolver==1.4.9
langdetect==1.0.9
lark==1.3.0
lazy_imports==1.0.1
llvmlite==0.45.0
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mistune==3.1.4
ml_dtypes==0.5.3
modelscope==1.32.0
more-itertools==10.8.0
mpmath==1.3.0
msoffcrypto-tool==5.4.2
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numba==0.62.0
numpy==2.2.6
olefile==0.47
ollama==0.6.0
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.22.1
openai==1.108.1
openai-whisper==20250625
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
paddle==1.1.10
paddleocr==3.3.2
paddlex==3.3.9
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
pdf2image==1.17.0
pdfminer.six==20250506
pexpect==4.9.0
pi_heif==1.1.0
pikepdf==9.11.0
pillow==11.3.0
platformdirs==4.4.0
posthog==6.7.5
prettytable==3.17.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydisort==1.3.6
Pygments==2.19.2
pyharp==1.7.8
pypandoc==1.15
pyparsing==3.2.5
pypdf==6.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
Send2Trash==1.8.3
sentence-transformers==5.1.0
shapely==2.1.2
six==1.17.0
snapy==0.8.6
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.5
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
tesseract==0.1.3
threadpoolctl==3.6.0
tiktoken==0.11.0
timm==1.0.20
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.7.1
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
unstructured==0.18.15
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wikipedia==1.4.0
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.9
